welt-home: Migracion de parches de datos d elos modulos
1- migramos y estructuramos el parche de insercion del footer y prefooter de una manera mas optima para hacer una insercion directa desde un solo parche llamado bloques.php AgSoftware_PageFooterV1
2- creamos una carpeta llamada html y se agrego el contenido de cada bloque es un archivo html para posterior mente llamarlo al archivo principal del parche AgSoftware_PageFooterV1.
3- migramos y estructuramos el parche de insercion del home de una manera mas optima para hacer una insercion directa desde un solo parche llamado bloques.php en el modulo AgSoftware_Home.
4- creamos una carpeta llamada html y se agrego el contenido de cada bloque es un archivo html para posterior mente llamarlo al archivo principal del parche AgSoftware_Home.
5- descomentar importacion del archivo de estilos _section-7-collapsible.less del archivo extend.less
6- agregar id a cada arreglo de dato del modulo AgSofware_Home.
