--- app/code/AgSoftware/Home/Setup/Patch/Data/AddBlockCmsAgileDevelopment.php
@@ -1,241 +1,0 @@
-<?php
-/**
- * Copyright © N/A All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace AgSoftware\Home\Setup\Patch\Data;
-
-use Magento\Eav\Setup\EavSetupFactory;
-use Magento\Framework\Setup\ModuleDataSetupInterface;
-use Magento\Framework\Setup\Patch\DataPatchInterface;
-use Magento\Framework\Setup\Patch\PatchRevertableInterface;
-use Magento\Framework\Stdlib\DateTime\Filter\Date;
-use Magento\Framework\Stdlib\DateTime\TimezoneInterface;
-use Magento\Framework\App\State;
-use Magento\Cms\Model\BlockFactory;
-use Magento\Framework\App\Area;
-
-class AddBlockCmsAgileDevelopment implements DataPatchInterface, PatchRevertableInterface 
-{
-
-    /**
-     * @var ModuleDataSetupInterface
-     */
-    private $moduleDataSetup;
-
-    /**
-     * Constructor
-     *
-     * @param ModuleDataSetupInterface $moduleDataSetup
-     * @param EavSetupFactory $eavSetupFactory
-     */
-    public function __construct (
-        Date $dateFilter,
-        TimezoneInterface $localeDate,
-        ModuleDataSetupInterface $moduleDataSetup,
-        State $state,
-        BlockFactory $cmsBlock,
-    ) {
-        $this->moduleDataSetup = $moduleDataSetup;
-        $this->cmsBlock = $cmsBlock;
-        $this->localeDate = $localeDate;
-        $this->_dateFilter = $dateFilter;
-
-        $this->state = $state;
-        //$this->cmsRepository = $cmsRepository;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function apply()
-    {
-        
-        $this->moduleDataSetup->getConnection()->startSetup();
-
-        /**
-         * @var \Magento\Cms\Model\Block $cmsBlock
-         */
-        $cmsBlock = $this->cmsBlock->create();
-        $data = [
-            "title" => "agile development",
-            "identifier" => "agile-development",
-            "store_id" => "All Store Views",
-            "content" => '<style>#html-body [data-pb-style=KESV8LO]{background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;align-self:stretch}#html-body [data-pb-style=DSQ387L]{display:flex;width:100%}#html-body [data-pb-style=SD4KR4E]{justify-content:flex-start;display:flex;flex-direction:column;background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;width:50%;align-self:stretch}#html-body [data-pb-style=LVSO4BR]{text-align:center;border-style:none}#html-body [data-pb-style=FFSTH73],#html-body [data-pb-style=XHCLQKA]{max-width:100%;height:auto}#html-body [data-pb-style=UO6HFCC]{justify-content:flex-start;display:flex;flex-direction:column;background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;width:50%;align-self:stretch}@media only screen and (max-width: 768px) { #html-body [data-pb-style=LVSO4BR]{border-style:none} }</style><div class="pagebuilder-column-group agile-development" data-background-images="{}" data-content-type="column-group" data-appearance="default" data-grid-size="12" data-element="main" data-pb-style="KESV8LO"><div class="pagebuilder-column-line" data-content-type="column-line" data-element="main" data-pb-style="DSQ387L"><div class="pagebuilder-column agile-development__image" data-content-type="column" data-appearance="full-height" data-background-images="{}" data-element="main" data-pb-style="SD4KR4E"><figure class="agile-development--imagen" data-content-type="image" data-appearance="full-width" data-element="main" data-pb-style="LVSO4BR"><img class="pagebuilder-mobile-hidden" src="{{media url=wysiwyg/Homepage-v2-ca.png}}" alt="" title="" data-element="desktop_image" data-pb-style="XHCLQKA"><img class="pagebuilder-mobile-only" src="{{media url=wysiwyg/Homepage-v2-ca.png}}" alt="" title="" data-element="mobile_image" data-pb-style="FFSTH73"></figure></div><div class="pagebuilder-column agile-development__content" data-content-type="column" data-appearance="full-height" data-background-images="{}" data-element="main" data-pb-style="UO6HFCC"><div class="agile-development__content--container" data-content-type="html" data-appearance="default" data-element="main">&lt;h2&gt;
-   AGILE DEVELOPMENT  
-   &lt;br&gt;
-   YOU CAN TRUST.
-&lt;/h2&gt;
-
-&lt;div id="accordion" data-mage-init='{
-        "accordion":{
-            "active": [0],
-            "collapsible": false,
-            "openedState": "active",
-            "multipleCollapsible": false,
-            "animate":{ "duration" :1000, "easing":"easeOutCubic"}
-        }}'&gt;
-    &lt;div data-role="collapsible" class="fist-item"&gt;
-        &lt;div data-role="trigger"&gt;
-            &lt;p&gt;Magento 1 certified and Magento 2 trained Developers&lt;/p&gt;
-        &lt;/div&gt;
-    &lt;/div&gt;
-    &lt;div data-role="content"&gt;
-      &lt;b&gt;
-         Our team throughout their career worked on many Magento Enterprise and Community Edition Clients. They implemented many complex solutions and integrated with ERP and Payment solutions.
-      &lt;/b&gt;
-      &lt;p&gt;
-         We are highly experienced, qualified and Magento Certified! For Magento 2 we went through the entire training program to get up to speed with the new infrastructure and best practices.
-      &lt;/p&gt;
-      &lt;p&gt;
-         While developing Magento 2 Theme Pearl, we had a phenomenal hands-on experience facing many challenges but we always managed to overcome!
-      &lt;/p&gt;
-    &lt;/div&gt;
-
-&lt;div data-role="collapsible"&gt;
-        &lt;div data-role="trigger"&gt;
-            &lt;p&gt;Project Management in Jira&lt;/p&gt;
-        &lt;/div&gt;
-    &lt;/div&gt;
-    &lt;div data-role="content"&gt;
-      &lt;b&gt;
-         The #1 software development tool used by agile teams
-      &lt;/b&gt;
-      &lt;p&gt;
-         We use Jira to track every project. No mater it&lsquo;s an internal project or your project. Staying organized and well planned out guarantees on-time delivery of every project.
-      &lt;/p&gt;
-    &lt;/div&gt;
-
-
-    &lt;div data-role="collapsible"&gt;
-        &lt;div data-role="trigger"&gt;
-            &lt;p&gt;Prototyping and design approval via InVision&lt;/p&gt;
-        &lt;/div&gt;
-    &lt;/div&gt;
-    &lt;div data-role="content"&gt; 
-      &lt;p&gt;
-         Do you have a specific design in your mind? We all do! Other companies will give you a simple Photoshop mock-up, a flat image that you cannot interact with.
-      &lt;/p&gt;
-      &lt;p&gt;
-         We do all our design in InVision! We add animations, gestures, and transitions to transform your static screens into clickable, interactive prototypes.
-      &lt;/p&gt;
-      &lt;p&gt;
-         You then give us your feedback directly on the design! Collaboration was never easier! 
-      &lt;/p&gt;
-      &lt;p&gt;
-         When you are happy with your design, we implement everything into your website. Your customers experience will be just like you imagined and it will interact as designed!
-      &lt;/p&gt;
-    &lt;/div&gt;
-
-&lt;div data-role="collapsible"&gt;
-        &lt;div data-role="trigger"&gt;
-            &lt;p&gt;Development Environment and Source Control in GitHub&lt;/p&gt;
-        &lt;/div&gt;
-    &lt;/div&gt;
-    &lt;div data-role="content"&gt;
-      &lt;p&gt;
-         Are you afraid that code changes or updates can destroy your website? We do too! Therefor the first thing we do is to setup a Development and Staging Environment! We test complex changes locally before deploying to staging or production using development tools like git and Vagrant.
-We use source control to make and track changes over time. Each branch is a separate environment.
-      &lt;/p&gt;
-    &lt;/div&gt;
-
-&lt;div data-role="collapsible"&gt;
-        &lt;div data-role="trigger"&gt;
-            &lt;p&gt;Cross-Device and Browser testing via BrowserStack&lt;/p&gt;
-        &lt;/div&gt;
-    &lt;/div&gt;
-    &lt;div data-role="content"&gt;
-      &lt;p&gt;
-         We want the websites that we develop to appear as they were designed to look on all browsers and all devices – whether that’s a desktop PC, Mac, Laptop, Tablet or Smartphone.
-      &lt;/p&gt;
-    &lt;/div&gt;
-
-&lt;div data-role="collapsible"&gt;
-        &lt;div data-role="trigger"&gt;
-            &lt;p&gt;We use PHPStorm to write all our code&lt;/p&gt;
-        &lt;/div&gt;
-    &lt;/div&gt;
-    &lt;div data-role="content"&gt;
-      &lt;p&gt;
-         Magento combines a lot of built-in tools and approaches designed to facilitate all development processes, and most significantly, enhance the source code quality. As a consequence, the product quality raises simultaneously.
-      &lt;/p&gt;
-      &lt;p&gt;
-         One of the most important tools for quality developments in Magento is PhpStorm:
-      &lt;/p&gt;
-      &lt;p class="margin-0"&gt;
-         ”Intelligent coding assistance”
-      &lt;/p&gt;
-      &lt;p class="margin-0"&gt;
-         ”PHP IDE that actually ‘gets’ your code”.
-      &lt;/p&gt;
-      &lt;p class="margin-0"&gt;
-         That’s what PhpStorm is for developers!
-      &lt;/p&gt;
-    &lt;/div&gt;
-
-&lt;div data-role="collapsible"&gt;
-        &lt;div data-role="trigger"&gt;
-            &lt;p&gt;Project collaboration and live chat via Slack&lt;/p&gt;
-        &lt;/div&gt;
-    &lt;/div&gt;
-    &lt;div data-role="content"&gt;
-      &lt;b&gt;
-         We love Slack! But really... who doesn&lsquo;t?
-      &lt;/b&gt;
-      &lt;p&gt;
-         We don&lsquo;t just use Slack for internal communication! We invite you to participate as well! We found that having a direct line to our developers and project managers makes everything much easier, faster and productive!
-      &lt;/p&gt;
-    &lt;/div&gt;
-
-&lt;/div&gt;</div></div></div></div>',
-            "is_active" => "1"
-        ];
-
-        $data["from_date"] = $this->localeDate->formatDate();
-        $filterValues = ["from_date" => $this->_dateFilter];
-
-        $inputFilter = new \Zend_Filter_Input(
-            $filterValues,
-            [],
-            $data
-        );
-        $data = $inputFilter->getUnescaped();
-        $cmsBlock->addData($data);
-
-        $this->state->setAreaCode(Area::AREA_ADMINHTML); // or \Magento\Framework\App\Area::AREA_FRONTEND, depending on your needs
-
-        $cmsBlock->save();
-
-        $this->moduleDataSetup->getConnection()->endSetup();
-
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function revert()
-    {
-        $this->moduleDataSetup->getConnection()->startSetup();
-
-        $this->moduleDataSetup->getConnection()->endSetup();
-    }
-
-    /**
-     * {@inheritdoc}
-    */
-    public function getAliases()
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-    */
-    public static function getDependencies()
-    {
-        return [];
-    }
-    
-}
