--- app/code/AgSoftware/Home/Setup/Patch/Data/AddBlockCmsAwesomeExtencions.php
@@ -1,145 +1,0 @@
-<?php
-/**
- * Copyright © N/A All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace AgSoftware\Home\Setup\Patch\Data;
-
-use Magento\Eav\Setup\EavSetupFactory;
-use Magento\Framework\Setup\ModuleDataSetupInterface;
-use Magento\Framework\Setup\Patch\DataPatchInterface;
-use Magento\Framework\Setup\Patch\PatchRevertableInterface;
-use Magento\Framework\Stdlib\DateTime\Filter\Date;
-use Magento\Framework\Stdlib\DateTime\TimezoneInterface;
-use Magento\Framework\App\State;
-use Magento\Cms\Model\BlockFactory;
-use Magento\Framework\App\Area;
-
-class AddBlockCmsAwesomeExtencions implements DataPatchInterface, PatchRevertableInterface 
-{
-
-    /**
-     * @var ModuleDataSetupInterface
-     */
-    private $moduleDataSetup;
-
-    /**
-     * Constructor
-     *
-     * @param ModuleDataSetupInterface $moduleDataSetup
-     * @param EavSetupFactory $eavSetupFactory
-     */
-    public function __construct (
-        Date $dateFilter,
-        TimezoneInterface $localeDate,
-        ModuleDataSetupInterface $moduleDataSetup,
-        State $state,
-        BlockFactory $cmsBlock,
-    ) {
-        $this->moduleDataSetup = $moduleDataSetup;
-        $this->cmsBlock = $cmsBlock;
-        $this->localeDate = $localeDate;
-        $this->_dateFilter = $dateFilter;
-
-        $this->state = $state;
-        //$this->cmsRepository = $cmsRepository;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function apply()
-    {
-        
-        $this->moduleDataSetup->getConnection()->startSetup();
-
-        /**
-         * @var \Magento\Cms\Model\Block $cmsBlock
-         */
-        $cmsBlock = $this->cmsBlock->create();
-        $data = [
-            "title" => "awesome extencsions",
-            "identifier" => "awesome-extencsions",
-            "store_id" => "All Store Views",
-            "content" => '<style>#html-body [data-pb-style=C8FFERQ],#html-body [data-pb-style=J6WHQHL]{background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll}#html-body [data-pb-style=J6WHQHL]{justify-content:flex-start;display:flex;flex-direction:column}#html-body [data-pb-style=C8FFERQ]{align-self:stretch}#html-body [data-pb-style=RLOB01G]{display:flex;width:100%}#html-body [data-pb-style=IBJ1U8W]{justify-content:flex-start;display:flex;flex-direction:column;width:100%}#html-body [data-pb-style=IBJ1U8W],#html-body [data-pb-style=VPXG53C]{background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;align-self:stretch}#html-body [data-pb-style=DV8J3LX]{display:flex;width:100%}#html-body [data-pb-style=RGR6G1T]{justify-content:flex-start;display:flex;flex-direction:column;background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;width:25%;align-self:stretch}#html-body [data-pb-style=N6858XK]{border-style:none}#html-body [data-pb-style=EYTJ5IS],#html-body [data-pb-style=TKB9JFP]{max-width:100%;height:auto}#html-body [data-pb-style=MY1KECN]{justify-content:flex-start;display:flex;flex-direction:column;background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;width:25%;align-self:stretch}#html-body [data-pb-style=KP66SO5]{border-style:none}#html-body [data-pb-style=JSAL70I],#html-body [data-pb-style=X5C7JNF]{max-width:100%;height:auto}#html-body [data-pb-style=CG6XSHE]{justify-content:flex-start;display:flex;flex-direction:column;background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;width:25%;align-self:stretch}#html-body [data-pb-style=AIGENV1]{border-style:none}#html-body [data-pb-style=IMQYN7D],#html-body [data-pb-style=L7C09AW]{max-width:100%;height:auto}#html-body [data-pb-style=V0Y8MB9]{justify-content:flex-start;display:flex;flex-direction:column;background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;width:25%;align-self:stretch}#html-body [data-pb-style=UQ2LPUF]{border-style:none}#html-body [data-pb-style=N3H9C8G],#html-body [data-pb-style=QEIG98D]{max-width:100%;height:auto}#html-body [data-pb-style=J85UV0X]{background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;align-self:stretch}#html-body [data-pb-style=C3TCLFP]{display:flex;width:100%}#html-body [data-pb-style=U26AV6G]{justify-content:flex-start;display:flex;flex-direction:column;background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;width:100%;align-self:stretch}@media only screen and (max-width: 768px) { #html-body [data-pb-style=AIGENV1],#html-body [data-pb-style=KP66SO5],#html-body [data-pb-style=N6858XK],#html-body [data-pb-style=UQ2LPUF]{border-style:none} }</style><div data-content-type="row" data-appearance="contained" data-element="main"><div class="awesome-extensions" data-enable-parallax="0" data-parallax-speed="0.5" data-background-images="{}" data-background-type="image" data-video-loop="true" data-video-play-only-visible="true" data-video-lazy-load="true" data-video-fallback-src="" data-element="inner" data-pb-style="J6WHQHL"><div class="pagebuilder-column-group awesome-extensions__wrapper" data-background-images="{}" data-content-type="column-group" data-appearance="default" data-grid-size="12" data-element="main" data-pb-style="C8FFERQ"><div class="pagebuilder-column-line" data-content-type="column-line" data-element="main" data-pb-style="RLOB01G"><div class="pagebuilder-column awesome-extencions__wrapper wrapper" data-content-type="column" data-appearance="full-height" data-background-images="{}" data-element="main" data-pb-style="IBJ1U8W"><div class="awesome-extensions--title wrapper--title" data-content-type="html" data-appearance="default" data-element="main">&lt;h2&gt;
-   WeltPixel Awesome
-   &lt;br&gt;
-   Extensions.
-&lt;/h2&gt;</div><div class="awesome-extensions--text wrapper--text wrapper--text-one" data-content-type="text" data-appearance="default" data-element="main"><p>Magento 2 modules developed by WeltPixel are compatible with both Magento 2 Commerce and Magento 2</p></div><div class="awesome-extensions--text wrapper--text wrapper--text-two" data-content-type="text" data-appearance="default" data-element="main"><p>Open Source. Extensions are used by thousands of users, were developed and tested with great care.</p></div></div></div></div><div class="pagebuilder-column-group awesome-extensions__sections" data-background-images="{}" data-content-type="column-group" data-appearance="default" data-grid-size="12" data-element="main" data-pb-style="VPXG53C"><div class="pagebuilder-column-line" data-content-type="column-line" data-element="main" data-pb-style="DV8J3LX"><div class="pagebuilder-column awesome-extensions__section-magento2 enhanced__quick-cart" data-content-type="column" data-appearance="full-height" data-background-images="{}" data-element="main" data-pb-style="RGR6G1T"><div class="awesome-extensions__sections-content enhanced__quick-cart__content" data-content-type="html" data-appearance="default" data-element="main">&lt;p class="awesome-extensions--span  enhanced__quick-cart--span"&gt;MAGENTO 2&lt;/p&gt;
-&lt;a hrfe="https://www.weltpixel.com/magento-2-quick-cart-checkout.html" class="awesome-extensions--title  enhanced__quick-cart--title"&gt;
-   &lt;h3&gt;ENHANCED&lt;br&gt;QUICK CART.&lt;/h3&gt;
-&lt;/a&gt;
-&lt;p class="awesome-extensions--text  enhanced__quick-cart--text"&gt;
-The modern responsive design combined with conversion oriented features offers the customer the ultimate shopping experience by allowing cart summary review in one glance without leaving the shopping page.
-&lt;/p&gt;</div><figure class="awesome-extencsions__sections-image enhanced__quick-cart__image" data-content-type="image" data-appearance="full-width" data-element="main" data-pb-style="N6858XK"><img class="pagebuilder-mobile-hidden" src="{{media url=wysiwyg/14-extensions-1.png}}" alt="" title="" data-element="desktop_image" data-pb-style="EYTJ5IS"><img class="pagebuilder-mobile-only" src="{{media url=wysiwyg/14-extensions-1.png}}" alt="" title="" data-element="mobile_image" data-pb-style="TKB9JFP"></figure></div><div class="pagebuilder-column awesome-extensions__section-magento2 ajax-search__autocomplete" data-content-type="column" data-appearance="full-height" data-background-images="{}" data-element="main" data-pb-style="MY1KECN"><div class="awesome-extensions__sections-content ajax-search__autocomplete__content" data-content-type="html" data-appearance="default" data-element="main">&lt;p class="awesome-extensions--span  ajax-search__autocomplete--span"&gt;MAGENTO 2&lt;/p&gt;
-&lt;a hrfe="https://www.weltpixel.com/magento-2-quick-cart-checkout.html" class="awesome-extensions--title  ajax-search__autocomplete--title"&gt;
-   &lt;h3&gt;AJAX SEARCH&lt;br&gt;AUTOCOMPLETE.&lt;/h3&gt;
-&lt;/a&gt;
-&lt;p class="awesome-extensions--text  ajax-search__autocomplete--text"&gt;
-Help your customers find the right products, increase conversion rate and reduce friction with ajax search autocomplete.
-&lt;/p&gt;</div><figure class="awesome-extensions__sections-image ajax-search__autocomplete__image" data-content-type="image" data-appearance="full-width" data-element="main" data-pb-style="KP66SO5"><img class="pagebuilder-mobile-hidden" src="{{media url=wysiwyg/14-extensions-2.png}}" alt="" title="" data-element="desktop_image" data-pb-style="JSAL70I"><img class="pagebuilder-mobile-only" src="{{media url=wysiwyg/14-extensions-2.png}}" alt="" title="" data-element="mobile_image" data-pb-style="X5C7JNF"></figure></div><div class="pagebuilder-column awesome-extensions__section-magento2 mega-menu" data-content-type="column" data-appearance="full-height" data-background-images="{}" data-element="main" data-pb-style="CG6XSHE"><div class="awesome-extensions__sections-content mega-menu__content" data-content-type="html" data-appearance="default" data-element="main">&lt;p class="awesome-extensions--span  mega-menu--span"&gt;MAGENTO 2&lt;/p&gt;
-&lt;a hrfe="https://www.weltpixel.com/magento-2-quick-cart-checkout.html" class="awesome-extensions--title  mega-menu--title"&gt;
-   &lt;h3&gt;MEGA&lt;br&gt;MENU.&lt;/h3&gt;
-&lt;/a&gt;
-&lt;p class="awesome-extensions--text  mega-menu--text"&gt;
-The Mega Menu extension helps merchants create a better look and professional navigation menu easily without technical skills. Categories can be presented in multiple formats in order to fit each category structure and provide a better navigation experience for the end user.
-&lt;/p&gt;</div><figure class="awesome-extensions__sections-image mega-menu__image" data-content-type="image" data-appearance="full-width" data-element="main" data-pb-style="AIGENV1"><img class="pagebuilder-mobile-hidden" src="{{media url=wysiwyg/14-extensions-3.png}}" alt="" title="" data-element="desktop_image" data-pb-style="L7C09AW"><img class="pagebuilder-mobile-only" src="{{media url=wysiwyg/14-extensions-3.png}}" alt="" title="" data-element="mobile_image" data-pb-style="IMQYN7D"></figure></div><div class="pagebuilder-column awesome-extensions__section-magento2 lazy-loading__enhanced" data-content-type="column" data-appearance="full-height" data-background-images="{}" data-element="main" data-pb-style="V0Y8MB9"><div class="awesome-extensions__sections-content lazy-loading__enhanced__content" data-content-type="html" data-appearance="default" data-element="main">&lt;p class="awesome-extensions--span  lazy-loading__enhanced--span"&gt;MAGENTO 2&lt;/p&gt;
-&lt;a hrfe="https://www.weltpixel.com/magento-2-quick-cart-checkout.html" class="awesome-extensions--title  lazy-loading__enhanced--title"&gt;
-   &lt;h3&gt;LAZY LOADING&lt;br&gt;ENHANCED.&lt;/h3&gt;
-&lt;/a&gt;
-&lt;p class="awesome-extensions--text  lazy-loading__enhancedt--text"&gt;
-The modern responsive design combined with conversion oriented features offers the customer the ultimate shopping experience by allowing cart summary review in one glance without leaving the shopping page.
-&lt;/p&gt;</div><figure class="awesome-extensions__sections-image lazy-loading__enhanced__image" data-content-type="image" data-appearance="full-width" data-element="main" data-pb-style="UQ2LPUF"><img class="pagebuilder-mobile-hidden" src="{{media url=wysiwyg/14-extensions-4.png}}" alt="" title="" data-element="desktop_image" data-pb-style="N3H9C8G"><img class="pagebuilder-mobile-only" src="{{media url=wysiwyg/14-extensions-4.png}}" alt="" title="" data-element="mobile_image" data-pb-style="QEIG98D"></figure></div></div></div><div class="awesome-extensions__container-button" data-content-type="html" data-appearance="default" data-element="main">&lt;a class="ancla-button awesome-extensions" href="https://www.weltpixel.com/magento2-extensions.html"&gt; 
-  &lt;button class="button-primary"&gt;VIEW ALL EXTENSIONS&lt;/button&gt;
- &lt;/a&gt;</div><div class="pagebuilder-column-group quote-4" data-background-images="{}" data-content-type="column-group" data-appearance="default" data-grid-size="12" data-element="main" data-pb-style="J85UV0X"><div class="pagebuilder-column-line" data-content-type="column-line" data-element="main" data-pb-style="C3TCLFP"><div class="pagebuilder-column quote-4__card" data-content-type="column" data-appearance="full-height" data-background-images="{}" data-element="main" data-pb-style="U26AV6G"><h2 class="quote-4--title" data-content-type="heading" data-appearance="default" data-element="main">WELTPIXEL STANDS OUT ABOVE THE REST</h2><div class="quote-4--review-title" data-content-type="text" data-appearance="default" data-element="main"><p>&nbsp;Google Analytics Enhanced Ecommerce for Magento 2.&nbsp;</p></div><div class="quote-4--text" data-content-type="text" data-appearance="default" data-element="main"><p>There are several extension for Tag Manager available for Magento 2. It was overwhelming figuring out which module is the best or the right one. They all looked good to me. Due to the feature set and asking around about the reputation of weltpixel, it was highly recommended that I use this solution. I installed it and I am impressed!!!! I&lsquo;ve used many modules in the past, but this one goes above and beyond. The developers really put some thought into their solution. They didn&lsquo;t just develop the extension, they have a full understanding of how Google Analytics and Tag Manager works. I didn&lsquo;t understand this about their solution until I installed it and was it magically configure all the things I could have easily spent days trying to figure out and tweak. Installing the module and putting in my account codes was all that was needed and the module did the rest. Kudo&lsquo;s - highly impressive solution!!!!!</p></div><div class="quote-4--name-author" data-content-type="text" data-appearance="default" data-element="main"><p>NLC</p></div></div></div></div></div></div>',
-            "is_active" => "1"
-        ];
-
-        $data["from_date"] = $this->localeDate->formatDate();
-        $filterValues = ["from_date" => $this->_dateFilter];
-
-        $inputFilter = new \Zend_Filter_Input(
-            $filterValues,
-            [],
-            $data
-        );
-        $data = $inputFilter->getUnescaped();
-        $cmsBlock->addData($data);
-
-        $this->state->setAreaCode(Area::AREA_ADMINHTML); // or \Magento\Framework\App\Area::AREA_FRONTEND, depending on your needs
-
-        $cmsBlock->save();
-
-        $this->moduleDataSetup->getConnection()->endSetup();
-
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function revert()
-    {
-        $this->moduleDataSetup->getConnection()->startSetup();
-
-        $this->moduleDataSetup->getConnection()->endSetup();
-    }
-
-    /**
-     * {@inheritdoc}
-    */
-    public function getAliases()
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-    */
-    public static function getDependencies()
-    {
-        return [];
-    }
-    
-}
