--- app/code/AgSoftware/Home/Setup/Patch/Data/AddBlockCmsPostList.php
@@ -1,139 +1,0 @@
-<?php
-/**
- * Copyright © N/A All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace AgSoftware\Home\Setup\Patch\Data;
-
-use Magento\Eav\Setup\EavSetupFactory;
-use Magento\Framework\Setup\ModuleDataSetupInterface;
-use Magento\Framework\Setup\Patch\DataPatchInterface;
-use Magento\Framework\Setup\Patch\PatchRevertableInterface;
-use Magento\Framework\Stdlib\DateTime\Filter\Date;
-use Magento\Framework\Stdlib\DateTime\TimezoneInterface;
-use Magento\Framework\App\State;
-use Magento\Cms\Model\BlockFactory;
-use Magento\Framework\App\Area;
-
-class AddBlockCmsPostList implements DataPatchInterface, PatchRevertableInterface 
-{
-
-    /**
-     * @var ModuleDataSetupInterface
-     */
-    private $moduleDataSetup;
-
-    /**
-     * Constructor
-     *
-     * @param ModuleDataSetupInterface $moduleDataSetup
-     * @param EavSetupFactory $eavSetupFactory
-     */
-    public function __construct (
-        Date $dateFilter,
-        TimezoneInterface $localeDate,
-        ModuleDataSetupInterface $moduleDataSetup,
-        State $state,
-        BlockFactory $cmsBlock,
-    ) {
-        $this->moduleDataSetup = $moduleDataSetup;
-        $this->cmsBlock = $cmsBlock;
-        $this->localeDate = $localeDate;
-        $this->_dateFilter = $dateFilter;
-
-        $this->state = $state;
-        //$this->cmsRepository = $cmsRepository;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function apply()
-    {
-        
-        $this->moduleDataSetup->getConnection()->startSetup();
-
-        /**
-         * @var \Magento\Cms\Model\Block $cmsBlock
-         */
-        $cmsBlock = $this->cmsBlock->create();
-        $data = [
-            "title" => "post list",
-            "identifier" => "post-list",
-            "store_id" => "All Store Views",
-            "content" => '<style>#html-body [data-pb-style=UNQTPFK]{background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;align-self:stretch}#html-body [data-pb-style=XM29HXU]{display:flex;width:100%}#html-body [data-pb-style=JJT2CF7]{justify-content:flex-start;display:flex;flex-direction:column;background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;width:33.3333%;align-self:stretch}#html-body [data-pb-style=L4DJ7L4]{text-align:center;border-style:none}#html-body [data-pb-style=AYUV2WO],#html-body [data-pb-style=FII4T84]{max-width:100%;height:auto}#html-body [data-pb-style=HU98YJ9]{justify-content:flex-start;display:flex;flex-direction:column;background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;width:33.3333%;align-self:stretch}#html-body [data-pb-style=SO2Q1FV]{text-align:center;border-style:none}#html-body [data-pb-style=SMG8DBS],#html-body [data-pb-style=SU3450Y]{max-width:100%;height:auto}#html-body [data-pb-style=JGC6FRB]{justify-content:flex-start;display:flex;flex-direction:column;background-position:left top;background-size:cover;background-repeat:no-repeat;background-attachment:scroll;width:33.3333%;align-self:stretch}#html-body [data-pb-style=MBATJLN]{text-align:center;border-style:none}#html-body [data-pb-style=RKA3H4R],#html-body [data-pb-style=SFWOX36]{max-width:100%;height:auto}@media only screen and (max-width: 768px) { #html-body [data-pb-style=L4DJ7L4],#html-body [data-pb-style=MBATJLN],#html-body [data-pb-style=SO2Q1FV]{border-style:none} }</style><div class="pagebuilder-column-group post-list" data-background-images="{}" data-content-type="column-group" data-appearance="default" data-grid-size="12" data-element="main" data-pb-style="UNQTPFK"><div class="pagebuilder-column-line" data-content-type="column-line" data-element="main" data-pb-style="XM29HXU"><div class="pagebuilder-column post-list__items server-side" data-content-type="column" data-appearance="full-height" data-background-images="{}" data-element="main" data-pb-style="JJT2CF7"><figure class="post-list--image server-side--image" data-content-type="image" data-appearance="full-width" data-element="main" data-pb-style="L4DJ7L4"><a href="https://www.weltpixel.com/blog/post/how-to-preview-and-debug-your-gtm-server-side-container" target="_blank" data-link-type="default" title="" data-element="link"><img class="pagebuilder-mobile-hidden" src="{{media url=wysiwyg/gtm-server-side-preview-debug.png}}" alt="" title="" data-element="desktop_image" data-pb-style="AYUV2WO"><img class="pagebuilder-mobile-only" src="{{media url=wysiwyg/gtm-server-side-preview-debug.png}}" alt="" title="" data-element="mobile_image" data-pb-style="FII4T84"></a></figure><div class="post-list--title server-side--title" data-content-type="html" data-appearance="default" data-element="main">&lt;h2&gt;
-   &lt;a href="https://www.weltpixel.com/blog/post/how-to-preview-and-debug-your-gtm-server-side-container"&gt; 
-      How To Preview and Debug Your GTM Server-Side Container?
-   &lt;/a&gt;
-&lt;/h2&gt;</div><div class="post-list--text server-side--text" data-content-type="html" data-appearance="default" data-element="main">&lt;a href="https://www.weltpixel.com/blog/post/how-to-preview-and-debug-your-gtm-server-side-container"&gt; 
-      &lt;p&gt;
-          It&lsquo;s time to verify and confirm that the data sent over from the client-side container is flowing into the server-side one. Read on to find out exactly how to do just that, step-by-step!
-      &lt;/p&gt;
-&lt;/a&gt;</div></div><div class="pagebuilder-column post-list__items google-analytics-4" data-content-type="column" data-appearance="full-height" data-background-images="{}" data-element="main" data-pb-style="HU98YJ9"><figure class="post-list--image google-analytics-4--image" data-content-type="image" data-appearance="full-width" data-element="main" data-pb-style="SO2Q1FV"><a href="https://www.weltpixel.com/blog/post/how-to-obtain-100-accurate-sales-data-in-google-analytics-4-ga4-via-measurement-protocol" target="_blank" data-link-type="default" title="" data-element="link"><img class="pagebuilder-mobile-hidden" src="{{media url=wysiwyg/ga4-measurement-protocol-main.png}}" alt="" title="" data-element="desktop_image" data-pb-style="SMG8DBS"><img class="pagebuilder-mobile-only" src="{{media url=wysiwyg/ga4-measurement-protocol-main.png}}" alt="" title="" data-element="mobile_image" data-pb-style="SU3450Y"></a></figure><div class="post-list--title google-analytics-4--title" data-content-type="html" data-appearance="default" data-element="main">&lt;h2&gt;
-   &lt;a href="https://www.weltpixel.com/blog/post/how-to-obtain-100-accurate-sales-data-in-google-analytics-4-ga4-via-measurement-protocol"&gt; 
-      How To Obtain 100% Accurate Sales Data in Google Analytics 4 (GA4) via Measurement Protocol
-   &lt;/a&gt;
-&lt;/h2&gt;</div><div class="post-list--text google-analytics-4--text" data-content-type="html" data-appearance="default" data-element="main">&lt;a href="https://www.weltpixel.com/blog/post/how-to-obtain-100-accurate-sales-data-in-google-analytics-4-ga4-via-measurement-protocol"&gt; 
-      &lt;p&gt;
-         Collecting data accurately via Google Analytics is becoming more and more difficult due to cookie restrictions and privacy settings. In this post, we&lsquo;ll look at how you can implement the Google Analytics Measurement Protocol to ensure your sales data is 100% accurate!
-      &lt;/p&gt;
-&lt;/a&gt;</div></div><div class="pagebuilder-column post-list__items product-update" data-content-type="column" data-appearance="full-height" data-background-images="{}" data-element="main" data-pb-style="JGC6FRB"><figure class="post-list--image product-update--image" data-content-type="image" data-appearance="full-width" data-element="main" data-pb-style="MBATJLN"><a href="https://www.weltpixel.com/blog/post/product-updates-and-new-releases-september-2022" target="_blank" data-link-type="default" title="" data-element="link"><img class="pagebuilder-mobile-hidden" src="{{media url=wysiwyg/sept-22-updates.png}}" alt="" title="" data-element="desktop_image" data-pb-style="SFWOX36"><img class="pagebuilder-mobile-only" src="{{media url=wysiwyg/sept-22-updates.png}}" alt="" title="" data-element="mobile_image" data-pb-style="RKA3H4R"></a></figure><div class="post-list--title product-update--title" data-content-type="html" data-appearance="default" data-element="main">&lt;h2&gt;
-   &lt;a href="https://www.weltpixel.com/blog/post/how-to-preview-and-debug-your-gtm-server-side-container"&gt; 
-      Product Updates and New Releases - September 2022
-   &lt;/a&gt;
-&lt;/h2&gt;</div><div class="post-list--text product-update--text" data-content-type="html" data-appearance="default" data-element="main">&lt;a href="https://www.weltpixel.com/blog/post/product-updates-and-new-releases-september-2022"&gt; 
-      &lt;p&gt;
-          Latest updates for WeltPixel Products, New Releases, and Change Logs. Release date September 1, 2022. 
-      &lt;/p&gt;
-&lt;/a&gt;</div></div></div></div>',
-            "is_active" => "1"
-        ];
-
-        $data["from_date"] = $this->localeDate->formatDate();
-        $filterValues = ["from_date" => $this->_dateFilter];
-
-        $inputFilter = new \Zend_Filter_Input(
-            $filterValues,
-            [],
-            $data
-        );
-        $data = $inputFilter->getUnescaped();
-        $cmsBlock->addData($data);
-
-        $this->state->setAreaCode(Area::AREA_ADMINHTML); // or \Magento\Framework\App\Area::AREA_FRONTEND, depending on your needs
-
-        $cmsBlock->save();
-
-        $this->moduleDataSetup->getConnection()->endSetup();
-
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function revert()
-    {
-        $this->moduleDataSetup->getConnection()->startSetup();
-
-        $this->moduleDataSetup->getConnection()->endSetup();
-    }
-
-    /**
-     * {@inheritdoc}
-    */
-    public function getAliases()
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-    */
-    public static function getDependencies()
-    {
-        return [];
-    }
-    
-}
