+++ app/code/AgSoftware/Home/Setup/Patch/Data/html/quote2.html
@@ -1,0 +1,91 @@
+<style>
+    #html-body [data-pb-style="DEC67PI"] {
+        background-position: left top;
+        background-size: cover;
+        background-repeat: no-repeat;
+        background-attachment: scroll;
+        text-align: center;
+        align-self: stretch;
+    }
+    #html-body [data-pb-style="ENO4DUE"] {
+        display: flex;
+        width: 100%;
+    }
+    #html-body [data-pb-style="Q7K32K6"] {
+        justify-content: flex-start;
+        display: flex;
+        flex-direction: column;
+        background-position: left top;
+        background-size: cover;
+        background-repeat: no-repeat;
+        background-attachment: scroll;
+        width: 100%;
+        align-self: stretch;
+    }
+</style>
+<div
+    class="pagebuilder-column-group quote-2"
+    data-background-images="{}"
+    data-content-type="column-group"
+    data-appearance="default"
+    data-grid-size="12"
+    data-element="main"
+    data-pb-style="DEC67PI"
+>
+    <div
+        class="pagebuilder-column-line"
+        data-content-type="column-line"
+        data-element="main"
+        data-pb-style="ENO4DUE"
+    >
+        <div
+            class="pagebuilder-column quote-2__card"
+            data-content-type="column"
+            data-appearance="full-height"
+            data-background-images="{}"
+            data-element="main"
+            data-pb-style="Q7K32K6"
+        >
+            <h3
+                class="quote-2--title"
+                data-content-type="heading"
+                data-appearance="default"
+                data-element="main"
+            >
+                INCREDIBLE CUSTOMER SERVICE
+            </h3>
+            <div
+                class="quote-2--review-title"
+                data-content-type="text"
+                data-appearance="default"
+                data-element="main"
+            >
+                <p>&nbsp;Pearl Theme for Magento 2.&nbsp;</p>
+            </div>
+            <div
+                class="quote-2--text"
+                data-content-type="text"
+                data-appearance="default"
+                data-element="main"
+            >
+                <p>
+                    The Welt Pixel team has been so helpful to us over the time
+                    that we&lsquo;ve been using their Pearl theme. The theme
+                    itself is really powerful and lets us do a lot more visually
+                    on our site than we ever considered before. We&lsquo;re
+                    hoping to move our secondary website to Magento in the near
+                    future and we&lsquo;ll definitely be using Welt Pixel when
+                    that happens. Thank you guys!!
+                </p>
+            </div>
+            <div
+                class="quote-2--name-author"
+                data-content-type="text"
+                data-appearance="default"
+                data-element="main"
+            >
+                <p>DEVIN</p>
+            </div>
+        </div>
+    </div>
+</div>
