+++ app/code/AgSoftware/Home/Setup/Patch/Data/html/themesAndExtencions.html
@@ -1,0 +1,113 @@
+<style>
+    #html-body [data-pb-style="R42VBCV"] {
+        background-position: left top;
+        background-size: cover;
+        background-repeat: no-repeat;
+        background-attachment: scroll;
+        align-self: stretch;
+    }
+    #html-body [data-pb-style="CRWC442"] {
+        display: flex;
+        width: 100%;
+    }
+    #html-body [data-pb-style="DECSX8G"],
+    #html-body [data-pb-style="GUARQWQ"] {
+        justify-content: flex-start;
+        display: flex;
+        flex-direction: column;
+        background-position: left top;
+        background-size: cover;
+        background-repeat: no-repeat;
+        background-attachment: scroll;
+        width: 50%;
+        align-self: stretch;
+    }
+    #html-body [data-pb-style="YEB0UBS"] {
+        border-style: none;
+    }
+    #html-body [data-pb-style="L664E3U"],
+    #html-body [data-pb-style="POYMJM7"] {
+        max-width: 100%;
+        height: auto;
+    }
+    @media only screen and (max-width: 768px) {
+        #html-body [data-pb-style="YEB0UBS"] {
+            border-style: none;
+        }
+    }
+</style>
+<div
+    class="pagebuilder-column-group themes-and-extencions"
+    data-background-images="{}"
+    data-content-type="column-group"
+    data-appearance="default"
+    data-grid-size="12"
+    data-element="main"
+    data-pb-style="R42VBCV"
+>
+    <div
+        class="pagebuilder-column-line"
+        data-content-type="column-line"
+        data-element="main"
+        data-pb-style="CRWC442"
+    >
+        <div
+            class="pagebuilder-column themes-and-extencions__content"
+            data-content-type="column"
+            data-appearance="full-height"
+            data-background-images="{}"
+            data-element="main"
+            data-pb-style="GUARQWQ"
+        >
+            <div
+                data-content-type="html"
+                data-appearance="default"
+                data-element="main"
+            >
+                &lt;div class="themes-and-extencions__content--title"&gt; &lt;h1
+                class="primary__title"&gt; magento themes and extencions.
+                &lt;/h1&gt; &lt;h3 class="secondary__title"&gt; Get straight to
+                growing your business! &lt;br&gt; From concept to launch in less
+                than 6 weeks. &lt;/h3&gt; &lt;/div&gt; &lt;div
+                class="themes-and-extencions__content--button"&gt; &lt;a
+                href="#" class="button__ancla"&gt; &lt;span
+                class="button__ancla--title"&gt;Explore Our
+                Solution&lt;/span&gt; &lt;span
+                class="button__ancla--text"&gt;Magento 2 Theme. Commerce &amp;
+                Open Source Ready&lt;/span&gt; &lt;/a&gt; &lt;/div&gt;
+            </div>
+        </div>
+        <div
+            class="pagebuilder-column"
+            data-content-type="column"
+            data-appearance="full-height"
+            data-background-images='{\"desktop_image\":\"{{media url=wysiwyg/laptop_banner1_1.png}}\",\"mobile_image\":\"{{media url=wysiwyg/hp-mob-1_1.png}}\"}'
+            data-element="main"
+            data-pb-style="DECSX8G"
+        >
+            <figure
+                class="themes-and-extencions__image"
+                data-content-type="image"
+                data-appearance="full-width"
+                data-element="main"
+                data-pb-style="YEB0UBS"
+            >
+                <img
+                    class="pagebuilder-mobile-hidden"
+                    src="{{media url=wysiwyg/laptop_banner1_4.png}}"
+                    alt=""
+                    title=""
+                    data-element="desktop_image"
+                    data-pb-style="L664E3U"
+                /><img
+                    class="pagebuilder-mobile-only"
+                    src="{{media url=wysiwyg/hp-mob-1_2.png}}"
+                    alt=""
+                    title=""
+                    data-element="mobile_image"
+                    data-pb-style="POYMJM7"
+                />
+            </figure>
+        </div>
+    </div>
+</div>
