+++ app/code/AgSoftware/PageFoteerV1/Setup/Patch/Data/html/preFooter.html
@@ -1,0 +1,36 @@
+<style>
+    #html-body [data-pb-style="S3UE4AA"] {
+        justify-content: flex-start;
+        display: flex;
+        flex-direction: column;
+        background-position: left top;
+        background-size: cover;
+        background-repeat: no-repeat;
+        background-attachment: scroll;
+    }
+</style>
+<div data-content-type="row" data-appearance="contained" data-element="main">
+    <div
+        class="pre-footer-container"
+        data-enable-parallax="0"
+        data-parallax-speed="0.5"
+        data-background-images="{}"
+        data-background-type="image"
+        data-video-loop="true"
+        data-video-play-only-visible="true"
+        data-video-lazy-load="true"
+        data-video-fallback-src=""
+        data-element="inner"
+        data-pb-style="S3UE4AA"
+    >
+        <div
+            class="pre-footer-container__img"
+            data-content-type="html"
+            data-appearance="default"
+            data-element="main"
+        >
+            &lt;img
+            src="https://www.weltpixel.com/media/wysiwyg/pearl-page/included_transparent_white_pre_footer.svg"&gt;
+        </div>
+    </div>
+</div>
