+++ app/design/frontend/Weltpixel/clone/web/css/source/_extend.less
@@ -21,7 +21,7 @@
 @import "./home/_program.less";
 @import "./home/_wlt-textcont.less";
 @import "./home/_content-items.less";
-/* @import "./home/_section-7-collapse.less"; */
+@import "./home/_section-7-collapse.less";
 @import "_footer.less";
